Added get_unit_id, set_unit_id getter/setter to pur_tcp_modbus. Not tested, but not used.
Also noticed that with 17.x eunit is failing on alternate runs because
of an address already in use error. Although this was known to happen
periodically in 18.x (due to what I believe are exit race conditions), the
17.x port fails consistently on back to back eunit runs. I believe this is due
to the replacement of gen_server:stop(...) in 18.x with
gen_server:call(kill, ...) in the 17.x port, not closing the associated server
side socket file descriptor properly. In the meantime, to get a successful run, wait between eunit executions. Will need to revisit this issue when I have time.